apps/ollam_ui/rend_backend/requirements.txt
fastapi
uvicorn
pydantic
duckduckgo-search>=6.3.0
httpx>=0.27.0
httpcore>=1.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
# Libraries for Code Interpreter
numpy
pandas
matplotlib
requests
pillow
scipy
scikit-learn